added deleteing files

+++ controllers/admin.js
@@ -1,4 +1,6 @@
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
+
 const { validationResult } = require("express-validator/check");
 
 exports.getAddProduct = (req, res, next) => {
@@ -153,7 +155,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = title;
       product.price = price;
       product.description = description;
-      if (image) product.imageUrl = image.path;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
 
       // need to nest this then here because this first if, don't want to continue
       return product.save().then(response => {
@@ -190,7 +195,12 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) return next(new Error("Product not found."));
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log(`Product with id ${prodId} removed`);
       res.redirect("/admin/products");
